+++ CHANGELOG.md
@@ -2,11 +2,18 @@
 All notable changes to this project will be documented in this file.
 
 
+## 0.2 - 2016-12-12
+### Fixed
+- Google Sheets updater
+### Added
+- commands
+### TODO
+- merge with remote-bms in raspberry
+
 ## 0.1 - 2016-12-12
 ### Added
 - gsheets module
 - create bot
-
 
 ## 0.0 - 2016-12-12
 ### Created